Move model stub to external class
